# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ACPBook Doc",
        "description": "Documentação api AcpBook",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Fazer login",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso. Retorna o token de autenticação.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Registrar novo usuário",
                "operationId": "8e4e3cb7b46126f4696379d3c7eeb4ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário registrado com sucesso. Retorna o token de autenticação.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Desconectar o usuário",
                "operationId": "ad65cbbd4e9f201619eae184a5961a98",
                "responses": {
                    "200": {
                        "description": "Desconectado com sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/post": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Lista de publicações",
                "operationId": "a8f335e0ced7872d7012086d0ecce9dd",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "ID do usuário para filtrar as publicações",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de publicações"
                    }
                }
            }
        },
        "/api/post/{id}/show": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Retornar postagem",
                "operationId": "7db240fd8f3c230b81bed8ceba264f78",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário para buscar as publicações",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da publicação"
                    }
                }
            }
        },
        "/api/post/store": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Criar nova publicação",
                "operationId": "f2e3743fe571eb83ec67d7be80533f2d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "description"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "ID do usuário",
                                        "type": "integer"
                                    },
                                    "content": {
                                        "description": "Conteúdo da publicação",
                                        "type": "string"
                                    },
                                    "tags": {
                                        "description": "Tags associadas ao post (não obrigatórias)",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Publicação criada com sucesso"
                    },
                    "400": {
                        "description": "Erro na requisição"
                    }
                }
            }
        },
        "/api/posts/{id}": {
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Atualizar publicação existente",
                "operationId": "650f4bed0846d9350903d8a17c7b503a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da publicação para ser atualizada",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "content"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "tags": {
                                        "description": "Tags associadas ao post (não obrigatórias)",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Publicação atualizada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Excluir publicação",
                "operationId": "62832788f1e230c55f7c8a9ec5b482fc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da publicação a ser excluída",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Publicação excluída com sucesso"
                    },
                    "404": {
                        "description": "Publicação não encontrada"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}